# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'wave-api API Documentation'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
...
paths:
...
  /api/users/nearby:
    get:
      summary: 'Get Nearby Users'
      operationId: getNearbyUsers
      description: "Retrieves users within a specified distance from the given latitude and longitude.\nIf latitude and longitude are not provided, the auth'd user's location will be used.\n\nthese are query params"
      parameters:
        -
          in: query
          name: latitude
          description: "The latitude of the user's location."
          example: 4326.41688
          required: false
          schema:
            type: number
            description: "The latitude of the user's location."
            example: 4326.41688
            nullable: false
        -
          in: query
          name: longitude
          description: "The longitude of the user's location."
          example: 4326.41688
          required: false
          schema:
            type: number
            description: "The longitude of the user's location."
            example: 4326.41688
            nullable: false
        -
          in: query
          name: distance
          description: 'The distance in kilometers to search for users.'
          example: 4326.41688
          required: true
          schema:
            type: number
            description: 'The distance in kilometers to search for users.'
            example: 4326.41688
            nullable: false
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n  \"users\": [\n    {id: \"98cca5ca-ca31-4031-a41b-241dc0876d5f\" name: \"John Doe\" distance: 1234},\n  ]"
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                latitude:
                  type: number
                  description: 'This field is required when <code>longitude</code> is present.'
                  example: 4326.41688
                  nullable: false
                longitude:
                  type: number
                  description: 'This field is required when <code>latitude</code> is present.'
                  example: 4326.41688
                  nullable: false
                distance:
                  type: number
                  description: 'Must be at least 1. Must not be greater than 1000.'
                  example: 17
                  nullable: false
              required:
                - distance
      security: []
...